tests/requirements.txt
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
kubernetes>=28.1.0
pyyaml>=6.0.1
requests>=2.31.0
colorama>=0.4.6
rich>=13.5.0

